namespace and time accumulation fixes
-Update internal PEG namespaces
-fixed numerous serious issues with accumulating timers that used floats. They now use doubles or avoid accumulation when possible.

+++ Runtime/CharacterControllerInterfaces.cs
@@ -52,7 +52,7 @@ namespace Peg.GCCI
     {
         bool GroundedEnabled { get; set; }
 
-        float LastGroundedTime { get; }
+        double LastGroundedTime { get; }
 
         /// <summary>
         /// Is this object stricktly grounded? Does not use additional fudge-factors such as timers or raycasts.
